+++ client/src/components/pages/Account/TeamPage.jsx
@@ -7,11 +7,13 @@ import BurgerMenuComp from "src/components/ui/Nav/BurgerMenuComp2";
 import bigCoinImage from "src/components/files/big-coin20.png";
 import ShopCard from "src/components/ui/Cards/ShopCard2";
 import axiosInstance from "src/axiosInstance";
+import { formatNumberWithSpaces } from 'src/utils';
 
 export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
   const [teamMembers, setTeamMembers] = useState([]);
   const [teamCost, setTeamCost] = useState(0);
-  const [result, setResult] = useState(null);
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  const [playerToSell, setPlayerToSell] = useState(null);
 
   useEffect(() => {
     if (!user) return;
@@ -26,6 +28,23 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
       });
   }, [user]);
 
+  const handleClickOutside = (event) => {
+    if (event.target.classList.contains(styles.overlay)) {
+      setIsMenuOpen(false);
+    }
+  };
+
+
+  const openSellMenu = (player) => {
+    setPlayerToSell(player);
+    setIsMenuOpen(true);
+  };
+
+  const closeSellMenu = () => {
+    setIsMenuOpen(false);
+    setPlayerToSell(null);
+  }
+
   function calculateTeamCost(teamArray) {
     let cost = 0;
     teamArray.forEach((teamPlayer) => {
@@ -34,10 +53,10 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
     setTeamCost(cost);
   }
 
-  async function sellPlayer(playerId, userId) {
+  async function sellPlayer() {
     try {
       const response = await axiosInstance.get(
-        `constract/checkPlayerInLiveRosters/${userId}/${playerId}`
+        `constract/checkPlayerInLiveRosters/${user.id}/${playerToSell.id}`
       );
       const result = response.data.result;
 
@@ -48,7 +67,7 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
       }
 
       const sellResponse = await axiosInstance.post(
-        `/players/sell/${playerId}/${userId}`
+        `/players/sell/${playerToSell.id}/${user.id}`
       );
       console.log(sellResponse.data);
       updateUserCoins(sellResponse.data);
@@ -66,6 +85,8 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
 
       alert(`Произошла ошибка при продаже игрока. ${errorMessage}`);
       console.error("Ошибка при продаже игрока:", error);
+    } finally {
+      closeSellMenu();
     }
   }
 
@@ -101,7 +122,7 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
                         user={user}
                         player={teamMember.player}
                         shop={false}
-                        sellPlayer={sellPlayer}
+                        openSellMenu={openSellMenu}
                       />
                     </div>
                   </div>
@@ -112,6 +133,34 @@ export default function TeamPage({ user, logoutHandler, updateUserCoins }) {
           </div>
         </div>
       </div>
+      {isMenuOpen && playerToSell && (
+               <div className={styles.overlay} onClick={handleClickOutside}>
+              <div className={styles.menu}>
+                <h3>Внимание!</h3>
+                <p>
+                  Вы подтверждаете продажу игрока <b>{playerToSell.nickname}</b>?
+                </p>
+                <p>
+                  Коммисия за продажу составит 2% —{" "}
+                  {formatNumberWithSpaces(
+                    Math.ceil((playerToSell.costcoins / 100) * 2)
+                  )}{" "}
+                  монет
+                </p>
+                <div className={styles.centerMode}>
+                  <button
+                    className={styles.menuButtonYes}
+                    onClick={sellPlayer}
+                  >
+                    Да
+                  </button>
+                  <button className={styles.menuButtonNo} onClick={closeSellMenu}>
+                    Нет
+                  </button>
+                </div>
+              </div>
+            </div>
+          )}
     </>
   );
 }
