+++ client/src/components/ui/Cards/Cards2.module.css
@@ -145,32 +145,6 @@
   font-family: "Montserrat", sans-serif;
 }
 
-
-
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5); /* Полупрозрачный чёрный */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000; /* Убедитесь, что меню поверх других элементов */
-}
-
-/* Стили для меню */
-.menu {
-  background-color: rgb(27, 27, 27);
-  padding: 30px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  position: relative;
-  width: 300px;
-  box-sizing: border-box;
-}
-
 .sellButton {
   background-color: #000;
   width: 144px;
@@ -204,41 +178,6 @@
   flex-direction: row;
   justify-content: center;
   align-items: center;
-}
-
-/* Стили для кнопки меню в всплывашке */
-.menuButtonYes {
-  padding: 10px 20px;
-  font-size: 16px;
-  cursor: pointer;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  width: 80px;
-}
-
-.menuButtonNo {
-  padding: 10px 20px;
-  font-size: 16px;
-  cursor: pointer;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  width: 80px;
-}
-
-.menuButtonNo:hover, .menuButtonYes:hover {
-  background-color: #1d14cd;
-}
-
-.centerMode {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 200px;
-  margin: 40px auto 10px;
 }
 
 h3 {
